# docs/openapi-spec.yaml
openapi: 3.0.3
info:
  title: NEXUS Platform API
  description: |
    # NEXUS Platform - Unified Backend API

    A comprehensive, AI-powered platform providing unified services for modern applications.

    ## Features

    * **SSOT (Single Source of Truth)**: Centralized configuration and alias management
    * **API Registry**: Service discovery and integration
    * **Rate Limiting**: Advanced rate limiting with multiple strategies
    * **Authentication**: JWT-based authentication with MFA support
    * **Financial Services**: Account management and transaction processing
    * **AI Services**: Chat, insights, and optimization features
    * **Monitoring**: Comprehensive system and performance monitoring
    * **Deployment**: Blue-green deployment and rollback capabilities

    ## Authentication

    Most endpoints require authentication. Include the JWT token in the Authorization header:

    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## Rate Limiting

    API requests are rate-limited. Check the response headers for limit information:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Time when limit resets

    ## Versioning

    API endpoints support versioning. Use the version prefix in the URL:
    - `v1`: Legacy endpoints
    - `v3`: Current stable version

    ## Error Responses

    All error responses follow this format:

    ```json
    {
...
      "details": {}
...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v3/auth/login:
    post:
      summary: User login
      description: Authenticate a user and return a JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "securePassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  status:
                    type: string
                    example: "success"
                  token:
                    type: string
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                  user:
                    type: string
                    example: "user@example.com"
                  expires_in:
                    type: integer
                    example: 3600
        '400':
          description: Invalid credentials
        '429':
          description: Rate limit exceeded
  /api/v3/auth/register:
    post:
      summary: User registration
      description: Create a new user account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "newuser@example.com"
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "SecurePass123"
                name:
                  type: string
                  example: "John Doe"
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
                  status:
                    type: string
                    example: "success"
                  user:
                    type: string
                    example: "newuser@example.com"
        '400':
          description: Invalid input or user already exists
        '429':
          description: Rate limit exceeded
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v3/auth/login
...